Save as JSON, CSV or Excel from Query Editor (#511)
Fixes #81. This initial implementation duplicates what is in the extension. When clicking the buttons, the user is prompted to type a file path, then we verify if it exists, save the contents and open the file. This also adds save as excel support that was included in the extension code.
TODO:

- Should add unit tests to cover this scenario
- Should use the file picker instead of asking to type a file path.