Save and Copy chart buttons (#1637)
Added buttons to save and copy chart data. Need a rethink of button look vs. icon, but this is in a state where it can be checked in and that part improved later.

Also fixed issue where calling SaveAsDialog broke mocha unit tests as it pulled in Electron dependencies which block test runs. This code path now has its own service that is usable by unit test code (as long as the service is mocked).

* Fixes #1640 cannot use IWindowIpcService in code that is unit tested