Fix connection dropping when zooming window (#3005)
- Fixes https://github.com/Microsoft/sqlopsstudio/issues/117
- Does not handle #3003 or #3004 which I discovered while testing this. These were already present, and I think we should just tackly in separate reviews.

The bug here was that the model save event is fired for any file, not just the active SQL one. So when settings were changed by zooming a save event occurrent and we tried to reroute the connection in the active editor to use the preferences URI instead (very bad). This also happened on saving a non-SQL file in a separate editor group.

The fix here was to move handling to the `saveAs` path only which is when a rename occurs. We hook into that action and find the affected SQL file & update the connection URI. This covers both untitled doc save and "Save As" for files on disk